+++ src/main/java/com/github/iridatelegrambot/command/HelpCommand.java
@@ -12,8 +12,9 @@ public class HelpCommand implements Command{
     public final static String HELP_MESSAGE = String.format("Доступные команды \n\n"
                                                 + "%s - начать работу со мной\n"
                                                 + "%s - приостановить работу\n"
-                                                + "%s - получить помощь\n",
-            START.getCommandName(),STOP.getCommandName(),HELP.getCommandName());
+                                                + "%s - получить помощь\n"
+                                                + "%s - получить информацию по работе бота\n",
+            START.getCommandName(),STOP.getCommandName(),HELP.getCommandName(),STAT.getCommandName());
 
     public HelpCommand(SendMessageServiceImpl sendMessage) {
         this.sendMessage = sendMessage;