+++ src/main/java/com/github/iridatelegrambot/command/StopCommand.java
@@ -1,20 +1,31 @@
 package com.github.iridatelegrambot.command;
 
 import com.github.iridatelegrambot.service.SendMessageServiceImpl;
+import com.github.iridatelegrambot.service.UserTelegramService;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
 public class StopCommand implements Command{
 
     private final SendMessageServiceImpl sendMessage;
+    private final UserTelegramService userTelegramService;
 
     public final static String STOP_MESSAGE = "Бот остановлен.";
 
-    public StopCommand(SendMessageServiceImpl sendMessage) {
+    public StopCommand(SendMessageServiceImpl sendMessage, UserTelegramService userTelegramService) {
         this.sendMessage = sendMessage;
+        this.userTelegramService = userTelegramService;
     }
 
     @Override
     public void execute(Update update) {
         sendMessage.sendMessage(update.getMessage().getChatId().toString(),STOP_MESSAGE);
+
+        Long chatId = update.getMessage().getChatId();
+
+        userTelegramService.findByChatId(chatId).ifPresent(
+                userTelegram -> {
+                    userTelegram.setActive(false);
+                    userTelegramService.save(userTelegram);
+                });
     }
 }