+++ src/main/java/com/github/iridatelegrambot/service/UserTelegramService.java
@@ -1,0 +1,17 @@
+package com.github.iridatelegrambot.service;
+
+import com.github.iridatelegrambot.entity.UserTelegram;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserTelegramService {
+
+    void save (UserTelegram user);
+
+    List<UserTelegram> getAllActiveUser();
+
+    List<UserTelegram> getAllUser();
+
+    Optional<UserTelegram> findByChatId(Long ChatId);
+}