+++ RELEASE_NOTES.md
@@ -1,9 +1,13 @@
 # Release Notes
 
+## 0.3.0-SNAPSHOT
+
+*   IRIDA-1: Added connection with the database and dockerfile.
+
 ## 0.2.0-SNAPSHOT
 
-*   IRIDA-2: Add telegram bot to the project
-*   IRIDA-3: Add first command for telegram bot
+*   IRIDA-2: Added telegram bot to the project
+*   IRIDA-3: Added first command for telegram bot
 
 ## 0.1.0-SNAPSHOT